requirement.txt
streamlit
pandas

openai
azure-storage-blob

tiktoken
scikit-learn
numpy

PyMuPDF
reportlab

aiohttp
requests

python-dateutil
